+++ Frends.FTP.UploadFiles/Frends.FTP.UploadFiles.Tests/Lib/Helpers.cs
@@ -1,7 +1,11 @@
 ﻿using FluentFTP;
 using Frends.FTP.UploadFiles.TaskConfiguration;
+using Frends.FTP.UploadFiles.Enums;
+using Ionic.Zip;
 using System;
+using System.Collections.Generic;
 using System.IO;
+using System.Text;
 
 namespace Frends.FTP.UploadFiles.Tests;
 
@@ -25,7 +29,9 @@ internal static class Helpers
             Password = FtpPassword,
             Port = FtpsPort,
             SslMode = FtpsSslMode.Explicit,
-            CertificateHashStringSHA1 = Sha1Hash
+            CertificateHashStringSHA1 = Sha1Hash,
+            BufferSize = 4096,
+            VerifyOption = VerifyOptions.None
         };
 
         return connection;
@@ -39,7 +45,9 @@ internal static class Helpers
             UserName = FtpUsername,
             Password = FtpPassword,
             Port = FtpPort,
-            SslMode = FtpsSslMode.None
+            SslMode = FtpsSslMode.None,
+            BufferSize = 4096,
+            VerifyOption = VerifyOptions.None
         };
 
         return connection;
@@ -49,18 +57,90 @@ internal static class Helpers
     {
         try
         {
-            var tmpFile = Path.GetTempFileName();
+            var tmpFile = Path.Combine(Path.GetDirectoryName(Path.GetTempFileName()), "test.zip");
             using (var client = new FtpClient(FtpHost, FtpPort, FtpUsername, FtpPassword))
             {
                 client.Connect();
                 client.SetWorkingDirectory(subDir);
                 client.DownloadFile(tmpFile, file);
             }
-            return File.ReadAllText(tmpFile);
+            return tmpFile;
         }
         catch (Exception ex)
         {
             throw new Exception(ex.Message);
+        }
+    }
+
+    internal static string CreateLargeDummyZipFiles(string dir, int count)
+    {
+        var name = "LargeTestFile";
+        var extension = ".bin";
+
+        var subDir = Path.Combine(dir, "zips");
+        Directory.CreateDirectory(subDir);
+
+        for (int i = 0; i < count; i++)
+        {
+            var path = Path.Combine(subDir, name + i + extension);
+            var fs = new FileStream(path, FileMode.CreateNew);
+            fs.Seek(2048L * 1024 * 512, SeekOrigin.Begin);
+            fs.WriteByte(0);
+            fs.Close();
+        }
+
+        using var zipFile = new Ionic.Zip.ZipFile(new UTF8Encoding(false));
+        zipFile.UseZip64WhenSaving = Zip64Option.Never;
+
+        foreach (var file in Directory.GetFiles(subDir))
+        {
+            zipFile.AddFile(file, "");
+        }
+
+        zipFile.Save(Path.Combine(dir, "test.zip"));
+
+        return Path.Combine(dir, "test.zip");
+    }
+
+    internal static bool ExtractLargeZipFile(string source, string destination)
+    {
+        try
+        {
+            var output = new UnzipOutput();
+            using var zip = ZipFile.Read(source);
+            string path = null;
+            zip.ExtractProgress += (sender, e) => Zip_ExtractProgress(e, output, path);
+            zip.ExtractExistingFile = ExtractExistingFileAction.OverwriteSilently;
+            zip.ExtractAll(destination);
+            return true;
+        }
+        catch
+        {
+            return false;
+        }
+    }
+
+    private static void Zip_ExtractProgress(ExtractProgressEventArgs e, UnzipOutput output, string fullPath)
+    {
+        if (e.EventType == ZipProgressEventType.Extracting_AfterExtractEntry && !e.CurrentEntry.IsDirectory)
+        {
+            // Path.GetFullPath changes directory separator to "\".
+            if (e.ExtractLocation == null) output.ExtractedFiles.Add(Path.GetFullPath(fullPath));
+            else output.ExtractedFiles.Add(Path.GetFullPath(Path.Combine(e.ExtractLocation, e.CurrentEntry.FileName)));
+        }
+    }
+
+    public class UnzipOutput
+    {
+        /// <summary>
+        /// a List-object of extracted files.
+        /// </summary>
+        /// <example>"ExtractedFiles": ["C:\\temp\\sample.txt",	"C:\\temp\\sample2.txt"]</example>
+        public List<string> ExtractedFiles { get; set; }
+
+        internal UnzipOutput()
+        {
+            ExtractedFiles = new List<string>();
         }
     }
 }